+++ docs/HISTORY.rst
@@ -8,3 +8,5 @@ Changelog
  - Add News, Story and Upate types
  - Add sinarngo.campaign relation behaviour
  - Add sinarngo.locations related country behavior
+ - Add initial view template
+ - Add source caption and url